Add files via upload
The client is a large MNC and they have 9 broad verticals across the organisation. One of the problem your client is facing is around identifying the right people for promotion (only for manager position and below) and prepare them in time. Currently the process, they are following is:
1. They first identify a set of employees based on recommendations/ past performance
2.  Selected employees go through the separate training and evaluation program for each vertical. These programs are based on the required skill of each vertical
3.  At the end of the program, based on various factors such as training performance, KPI completion (only employees with KPIs completed greater than 60% are considered) etc., employee gets promotion.
For above mentioned process, the final promotions are only announced after the evaluation and this leads to delay in transition to their new roles. Hence, company needs your help in identifying the eligible candidates at a particular checkpoint so that they can expedite the entire promotion cycle. 
They have provided multiple attributes around Employee's past and current performance along with demographics. Now, the task is to predict whether a potential promotee at checkpoint in the test set will be promoted or not after the evaluation process.

The columns having 1 and 0 as values are represented as below:
-if Percent of KPIs(Key performance Indicators) >80% then 1 else 0
-if awards won during previous year then 1 else 0

Expectations and Evaluation:
We expect the candidate to build a model upon training data (70% of the train_hr.csv) and evaluate the model on the validation data, (30% of the train_hr.csv). Use the random_state=10 while splitting the data. Use the model to predict upon the test_hr.csv. 

I have used the Logistic Regression approach to build and evaluate the model.